requirements.txt
typer==0.12.5
pydantic==2.9.2
PyYAML==6.0.2
feedparser==6.0.11
beautifulsoup4==4.12.3
httpx==0.27.2
tenacity==9.0.0
python-dateutil==2.9.0.post0
pandas==2.2.3
duckdb==1.1.3
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
sqlite-utils==3.36
uvloop==0.20.0; platform_system != "Windows"